KTOR-4960 Make sessions plugin and auth provider multiplatform (#3187)
