+++ ktor-server/build.gradle.kts
@@ -3,7 +3,6 @@ description = "Wrapper for ktor-server-core and base plugins"
 kotlin.sourceSets {
     jvmMain {
         dependencies {
-            api(project(":ktor-server:ktor-server-plugins:ktor-server-sessions"))
             api(project(":ktor-server:ktor-server-plugins:ktor-server-call-logging"))
             api(project(":ktor-server:ktor-server-plugins:ktor-server-default-headers"))
             api(project(":ktor-server:ktor-server-plugins:ktor-server-compression"))
@@ -26,6 +25,7 @@ kotlin.sourceSets {
             api(project(":ktor-server:ktor-server-plugins:ktor-server-partial-content"))
             api(project(":ktor-server:ktor-server-plugins:ktor-server-status-pages"))
             api(project(":ktor-server:ktor-server-plugins:ktor-server-method-override"))
+            api(project(":ktor-server:ktor-server-plugins:ktor-server-sessions"))
         }
     }
 }
