+++ ktor-server/ktor-server-plugins/ktor-server-auth/jvmAndNix/test/io/ktor/tests/auth/SessionAuthTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2021 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2014-2022 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
 package io.ktor.tests.auth
@@ -14,7 +14,9 @@ import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import io.ktor.server.sessions.*
 import io.ktor.server.testing.*
+import io.ktor.utils.io.core.*
 import kotlinx.coroutines.*
+import kotlinx.serialization.*
 import kotlin.test.*
 
 @Suppress("DEPRECATION")
@@ -123,5 +125,6 @@ class SessionAuthTest {
         }
     }
 
+    @Serializable
     data class MySession(val id: Int) : Principal
 }
