+++ ktor-server/ktor-server-plugins/ktor-server-sessions/build.gradle.kts
@@ -10,9 +10,10 @@ plugins {
 
 kotlin {
     sourceSets {
-        jvmMain {
+        jvmAndNixMain {
             dependencies {
                 api(libs.kotlinx.serialization.core)
+                api(libs.kotlinx.serialization.json)
             }
         }
         jvmTest {
