+++ ktor-server/ktor-server-plugins/ktor-server-sessions/jvmAndNix/test/io/ktor/server/sessions/SessionTransportCookieTest.kt
@@ -2,12 +2,9 @@
  * Copyright 2014-2022 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package io.ktor.tests.sessions
+package io.ktor.server.sessions
 
-import io.ktor.server.sessions.*
-import org.junit.Test
 import kotlin.test.*
-import kotlin.time.*
 
 class SessionTransportCookieTest {
