+++ ktor-utils/js/src/io/ktor/util/collections/ConcurrentMapJs.kt
@@ -53,6 +53,12 @@ public actual class ConcurrentMap<Key, Value> public actual constructor(initialC
 
     override fun remove(key: Key): Value? = delegate.remove(key)
 
+    public actual fun remove(key: Key, value: Value): Boolean {
+        if (delegate[key] != value) return false
+        delegate.remove(key)
+        return true
+    }
+
     override fun hashCode(): Int = delegate.hashCode()
 
     override fun equals(other: Any?): Boolean {
